libvpx: fix compile error on Mt. Lion
libvpx-1.1.0 has a problem on Mt. Lion where an asm file is not
created due to an extra operator in a regular expression, where
the file `asm_enc_offsets.asm` goes missing.

Add a patch to fix the regular expression.  See the issue and
formula comments for links to the analysis.

Remove the unneeded `--disable-debug` and `--enable-vp8` flags
because they are the default.

Tested on Snow Leopard, Lion, and Mt. Lion using XCode-4.0.2, 4.3.3,
and 4.5.

Fixes #12567

Signed-off-by: Adam Vandenberg <flangy@gmail.com>